# specs/core/openapi.yaml
openapi: 3.0.3
info:
  title: AGRO MIRAI API
  version: v1
  description: |
    Contract skeleton for the AGRO MIRAI HTTP API. This document defines
    the shape every endpoint will honour once Module 11 implements them;
    the frontend (Module 14) may type against it beforehand.

    Conventions: see docs/conventions.md. Enums: see specs/core/enums.md.
    Entity schemas mirror specs/core/schema.yaml.
servers:
  - url: https://api.agromirai.local/v1
    description: Placeholder; real hosts land in Module 11.
security:
  - bearerAuth: []
paths:
...
  /farmers/me:
    get:
      summary: Get the authenticated farmer's profile.
      responses:
        '200':
          description: Farmer profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Farmer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /fields:
    get:
      summary: List Fields owned by the authenticated Farmer.
      responses:
        '200':
          description: A list of Fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Field' }
                required: [items]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a Field.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FieldCreate' }
      responses:
        '201':
          description: The created Field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Field' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /fields/{field_id}:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get a Field by id.
      responses:
        '200':
          description: The Field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Field' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v2/farmers/me:
    get:
      summary: Get the session-authenticated Farmer's own profile.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: Farmer profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Farmer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      summary: Update the session-authenticated Farmer's own profile.
      description: |
        Partial update — only keys present in the body are changed.
        Mutable fields: ``name``, ``preferred_language``, ``photo_url``.
        Email and password changes are out of scope for v1 (see ADR 0017 known limitations).
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                preferred_language:
                  type: string
                  enum: [en, kn, te, hi]
                  description: Must be one of V1_LANGUAGES — the languages the voice stack can act on (Module 25).
                photo_url:
                  type: string
                  description: Module 30. A base64 data URI (e.g. "data:image/jpeg;base64,..."). Small images only, downscaled/compressed client-side.
              minProperties: 1
      responses:
        '200':
          description: Updated Farmer profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Farmer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v2/fields:
    get:
      summary: List Fields owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: A list of Fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Field' }
                required: [items]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a Field owned by the session-authenticated Farmer.
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FieldCreate' }
      responses:
        '201':
          description: The created Field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Field' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v2/fields/{field_id}:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: Get a Field by id. 404 (not 403) if it exists but is owned by a different Farmer — no cross-tenant existence leak.
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: The Field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Field' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: |
        Module 23 (A5). Partial update of a Field owned by the
        session-authenticated Farmer — only the keys present in the body
        are changed. /v1 deliberately does not get this route (frozen
        per ADR 0017); no /v1 caller has ever needed field mutation.
      security: [ { sessionCookie: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FieldUpdate' }
      responses:
        '200':
          description: The updated Field.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Field' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: |
        Delete a Field owned by the session-authenticated Farmer.
        All linked data (advisories, soil samples, weather readings, NDVI
        readings, irrigation advice, disease risk alerts, feedback) is
        cascade-deleted per the SQLite/Postgres migration's ON DELETE CASCADE.
        Returns 204 on success, 404 if not found or owned by a different Farmer.
      security: [ { sessionCookie: [] } ]
      responses:
        '204':
          description: Field deleted successfully. No response body.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- v2: value endpoints + voice (Module 23) ---
  # Blocker A: everything below /v2/fields/{field_id}/... and /v2/feedback
  # carries the product's actual value and previously existed only under
  # /v1's single shared FARMER_ID — these are the session-authenticated
  # equivalents, sharing their handler bodies with /v1 (never their route
  # registration) via agro_mirai.api.value_endpoints — see
  # decisions/0020-v2-value-endpoints-and-voice-api.md.
  /v2/fields/{field_id}/refresh-data:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    post:
      summary: >-
        Module 34 follow-up 2: manually re-pull real weather/soil/NDVI
        data for a Field owned by the session-authenticated Farmer via
        the Module 03 acquisition adapters, the same call field creation
        makes. Weather/soil are fetched synchronously; NDVI is fetched in
        a background thread and is not reflected in this response
        (data_acquisition.ndvi is always "gathering"). No automatic
        periodic refresh exists — this manual trigger is the deliberate
        MVP, not a placeholder for a scheduler.
      security: [ { sessionCookie: [] } ]
      responses:
        '202':
          description: Refresh accepted; weather/soil status reflects the synchronous portion, NDVI is still in flight.
          content:
            application/json:
              schema:
                type: object
                properties:
                  field_id: { type: string }
                  data_acquisition:
                    type: object
                    properties:
                      weather: { type: string, enum: [ready, unavailable] }
                      soil: { type: string, enum: [ready, unavailable] }
                      ndvi: { type: string, enum: [gathering] }
                    required: [weather, soil, ndvi]
                required: [field_id, data_acquisition]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v2/fields/{field_id}/data-summary:
    parameters:
      - $ref: '#/components/parameters/FieldId'
    get:
      summary: >-
        Module 39 (additive): the real weather / soil / NDVI rows already
        persisted for a Field owned by the session-authenticated Farmer, for
        display. weather.current is the newest observed reading, recent the
        last 7 observed, forecast up to 7 upcoming forecast rows. soil and
        ndvi.latest are null until that source has produced data (NDVI is
        fetched in a background thread after field creation). Each row
        carries its own `source` string (e.g. open-meteo, soilgrids, gee).
      security: [ { sessionCookie: [] } ]
      responses:
        '200':
          description: Data summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  field_id: { type: string }
                  fetched_at: { type: string, format: date-time }
                  soil_used: { type: object, nullable: true, additionalProperties: true }
                  weather:
                    type: object
                    properties:
                      current: { nullable: true, $ref: '#/components/schemas/WeatherReading' }
                      recent: { type: array, items: { $ref: '#/components/schemas/WeatherReading' } }
                      forecast: { type: array, items: { $ref: '#/components/schemas/WeatherReading' } }
                  soil: { nullable: true, $ref: '#/components/schemas/SoilSample' }
                  ndvi:
                    type: object
                    properties:
                      latest: { nullable: true, $ref: '#/components/schemas/NDVIReading' }
                      history: { type: array, items: { $ref: '#/components/schemas/NDVIReading' } }
                required: [field_id, weather, soil, ndvi]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask signed session cookie issued by POST /v2/auth/login (Module 19). Not a bearer token — the cookie is opaque and itsdangerous-signed.
  parameters:
    FieldId:
      name: field_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found or not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    WeatherReading:
      type: object
      required: [id, field_id, observed_at, source, temp_c, is_forecast]
      properties:
        id: { type: string }
        field_id: { type: string }
        observed_at: { type: string, format: date-time }
        source: { type: string }
        temp_c: { type: number }
        temp_min_c: { type: number, nullable: true }
        temp_max_c: { type: number, nullable: true }
        humidity_pct: { type: number, nullable: true }
        rainfall_mm: { type: number, nullable: true }
        wind_mps: { type: number, nullable: true }
        is_forecast: { type: boolean }
    SoilSample:
      type: object
      required: [id, field_id, observed_at, source]
      properties:
        id: { type: string }
        field_id: { type: string }
        observed_at: { type: string, format: date-time }
        source: { type: string }
        ph: { type: number, nullable: true }
        nitrogen_mg_per_kg: { type: number, nullable: true }
        phosphorus_mg_per_kg: { type: number, nullable: true }
        potassium_mg_per_kg: { type: number, nullable: true }
        organic_carbon_pct: { type: number, nullable: true }
        moisture_pct: { type: number, nullable: true }
    NDVIReading:
      type: object
      required: [id, field_id, observed_at, source, ndvi]
      properties:
        id: { type: string }
        field_id: { type: string }
        observed_at: { type: string, format: date-time }
        source: { type: string }
        ndvi: { type: number }
        cloud_cover_pct: { type: number, nullable: true }
        satellite: { type: string, nullable: true }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Farmer:
      type: object
      required: [id, created_at, updated_at, name, preferred_language]
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        name: { type: string }
        preferred_language:
          type: string
          enum: [en, hi, kn, ta, te, mr, bn, gu]
        phone: { type: string, nullable: true }
        district: { type: string, nullable: true }
        state: { type: string, nullable: true }
        email:
          type: string
          format: email
          nullable: true
          description: Module 19. Present for farmers registered via /v2. Never accompanied by password_hash in a response.
        role:
          type: string
          enum: [farmer, admin]
          nullable: true
          description: Module 19. Absent/defaults to "farmer" for pre-Module-19 records.
        photo_url:
          type: string
          nullable: true
          description: Module 30. A base64 data URI -- no file-storage service exists yet, so the profile photo is stored directly rather than as a link to one.
...
    Field:
      type: object
      required: [id, farmer_id, created_at, updated_at, name, latitude, longitude, area_ha]
      properties:
        id: { type: string, format: uuid }
        farmer_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        name: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        area_ha: { type: number, minimum: 0 }
        elevation_m: { type: number, nullable: true }
        soil_type:
          type: string
          nullable: true
          enum: [alluvial, black, red, laterite, mountain, desert, saline, peaty, unknown]
        current_crop:
          type: string
          nullable: true
          enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
        sown_on: { type: string, format: date, nullable: true }
    FieldCreate:
      type: object
      required: [name, latitude, longitude, area_ha]
      properties:
        name: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        area_ha: { type: number, minimum: 0 }
        elevation_m: { type: number }
        soil_type:
          type: string
          enum: [alluvial, black, red, laterite, mountain, desert, saline, peaty, unknown]
        current_crop:
          type: string
          enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
        sown_on: { type: string, format: date }
...
    FieldUpdate:
      description: Module 23 (A5). Every property optional (partial update); at least one must be present.
      type: object
      minProperties: 1
      properties:
        name: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        area_ha: { type: number, minimum: 0 }
        elevation_m: { type: number, nullable: true }
        soil_type:
          type: string
          nullable: true
          enum: [alluvial, black, red, laterite, mountain, desert, saline, peaty, unknown]
        current_crop:
          type: string
          nullable: true
          enum: [rice, maize, chickpea, kidneybeans, pigeonpeas, mothbeans, mungbean, blackgram, lentil, pomegranate, banana, mango, grapes, watermelon, muskmelon, apple, orange, papaya, coconut, cotton, jute, coffee]
        sown_on: { type: string, format: date, nullable: true }
...